MWC/requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.24.0

# LoRA and Quantization
peft>=0.6.0
bitsandbytes>=0.41.0

# Audio Processing
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1
audioread>=3.0.0

# Evaluation Metrics
sacrebleu>=2.3.1
jiwer>=3.0.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.16.0
onnxruntime-tools>=1.7.0

# Utilities
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Optional: TensorBoard for monitoring
tensorboard>=2.14.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0
